+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [DesignToolbox] Files for agentic AI (Orange-OpenSource/ouds-ios#1341)
 - [DesignToolbox] In the design toolbox, use grid margin helper (Orange-OpenSource/ouds-ios#1299)
+- [Library] Update Bullet list API to rename unordered icon to asset (Orange-OpenSource/ouds-ios#1326)
+- [DesignToolbox] Use of updated `bullet list` component API ("unordered icon" to "unordered asset") (Orange-OpenSource/ouds-ios#1326)
 - [Library] Signatures of control-item-based components (Orange-OpenSource/ouds-ios#1314)
 - [Library] `inline alert` component (Orange-OpenSource/ouds-ios#1307)
 - [Library] `alert message` component (Orange-OpenSource/ouds-ios#1159)